Made some changes to add differentiation to accounts
