+++ app/src/main/java/com/example/brett/landlord_tenant/MainActivity.java
@@ -1,6 +1,9 @@
 package com.example.brett.landlord_tenant;
 
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.InputType;
@@ -18,16 +21,32 @@ import org.w3c.dom.Text;
 
 import java.util.HashMap;
 
-public class TenantMainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity {
 
     String name = "";
     String identifier = "";
     String username = "";
     String landlordName ="";
+
+    private BroadcastReceiver broadcastReceiver;
+    @Override
+    protected void onResume(){
+        super.onResume();
+
+        if(broadcastReceiver == null){
+            broadcastReceiver = new BroadcastReceiver() {
+                @Override
+                public void onReceive(Context context, Intent intent) {
+
+                }
+            };
+        }
+        registerReceiver(broadcastReceiver, new IntentFilter("automatic_checkIn"));
+
+    }
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_tenant_main);
         Intent intent = getIntent();
         Bundle extras = intent.getExtras();
         if(extras!= null){
@@ -35,6 +54,17 @@ public class TenantMainActivity extends AppCompatActivity {
             identifier = extras.getString("identifier");
             username = extras.getString("username");
         }
+
+        if(identifier.equals("landlord")){
+            setContentView(R.layout.activity_landlord_main);
+        }
+        else if (identifier.equals("tenant")){
+            setContentView(R.layout.activity_tenant_main);
+        }
+        else{
+            setContentView(R.layout.activity_tenant_main);
+        }
+
         TextView welcome_name = (TextView) findViewById(R.id.welcome_name);
         welcome_name.setText(name);
 
@@ -88,16 +118,22 @@ public class TenantMainActivity extends AppCompatActivity {
 
     public void contacts(View view){
         Intent intent = new Intent(this, contactActivity.class);
+        intent.putExtra("name", name);
+        intent.putExtra("identifier", identifier);
         startActivity(intent);
     }
 
     public void utilities(View view){
         Intent intent = new Intent(this, utilitiesActivity.class);
+        intent.putExtra("name", name);
+        intent.putExtra("identifier", identifier);
         startActivity(intent);
     }
 
     public void lease(View view){
         Intent intent = new Intent(this, leaseActivity.class);
+        intent.putExtra("name", name);
+        intent.putExtra("identifier", identifier);
         startActivity(intent);
     }
 }
